% tinyrox says don't edit this manually, but it can't stop you!
\name{check_code_lines}
\alias{check_code_lines}
\title{Check Code Lines for Issues}
\usage{
check_code_lines(lines, filename)
}
\arguments{
\item{lines}{Character vector of code lines}

\item{filename}{Filename for reporting}
}
\value{
List of issues
}
\description{
Check Code Lines for Issues
}
