% Generated by rhydrogen: do not edit by hand
% Please edit documentation in R/release.R
\name{check_win_devel}
\alias{check_win_devel}
\title{Check Package on Windows via win-builder}
\usage{
check_win_devel(path = ".", r_version = c("devel", "release", "oldrelease"))
}
\arguments{
\item{path}{Path to package root directory.}

\item{r_version}{Which R version to test: "devel", "release", or "oldrelease". Default is
"devel".}
}
\value{
TRUE if upload succeeded (invisibly).
}
\description{
Uploads package to win-builder.r-project.org for testing on Windows.
Results are emailed to the package pkg_maintainer.
}
\examples{
\dontrun{
check_win_devel()
check_win_devel(r_version = "release")
}
}
