% Generated by rhydrogen: do not edit by hand
% Please edit documentation in R/rd.R
\name{escape_rd}
\alias{escape_rd}
\title{Escape Special Characters for Rd}
\description{Escape Special Characters for Rd}
\usage{
escape_rd(text)
}
\arguments{
  \item{text}{Text to escape.}
}
\value{
Escaped text.
}
\keyword{internal}
