% tinyrox says don't edit this manually, but it can't stop you!
\name{escape_regex}
\alias{escape_regex}
\title{Escape Regex Special Characters}
\usage{
escape_regex(x)
}
\arguments{
\item{x}{String to escape}
}
\value{
Escaped string safe for use in regex
}
\description{
Escape Regex Special Characters
}
