% tinyrox says don't edit this manually, but it can't stop you!
\name{extract_formals}
\alias{extract_formals}
\title{Extract Function Formals from Code}
\usage{
extract_formals(code)
}
\arguments{
\item{code}{Code starting with "function("}
}
\value{
List with 'names' (argument names) and 'usage' (formatted for Rd).
}
\description{
Extract Function Formals from Code
}
\keyword{internal}
