% tinyrox says don't edit this manually, but it can't stop you!
\name{extract_function_name}
\alias{extract_function_name}
\title{Extract Function Name from Definition Line}
\usage{
extract_function_name(line)
}
\arguments{
\item{line}{Code line potentially containing function definition}
}
\value{
Function name or NULL
}
\description{
Extract Function Name from Definition Line
}
