% tinyrox says don't edit this manually, but it can't stop you!
\name{find_unquoted_names}
\alias{find_unquoted_names}
\title{Find Unquoted Names in Text}
\usage{
find_unquoted_names(text, names)
}
\arguments{
\item{text}{Text to search}

\item{names}{Names to look for}
}
\value{
Character vector of unquoted names found
}
\description{
Finds package/software names that appear without single quotes.
}
