% Generated by rhydrogen: do not edit by hand
% Please edit documentation in R/rd.R
\name{get_package_name}
\alias{get_package_name}
\title{Get Package Name from DESCRIPTION}
\usage{
get_package_name(path)
}
\arguments{
\item{path}{Package root path.}
}
\value{
Package name.
}
\description{
Get Package Name from DESCRIPTION
}
\keyword{internal}
