% tinyrox says don't edit this manually, but it can't stop you!
\docType{data}
\name{KNOWN_S3_GENERICS}
\alias{KNOWN_S3_GENERICS}
\title{Known S3 Generic Functions}
\format{
An object of class \code{character} of length 139.
}
\usage{
KNOWN_S3_GENERICS
}
\description{
List of base R S3 generics for auto-detection when @export is used.
}
\keyword{internal}
