% tinyrox says don't edit this manually, but it can't stop you!
\name{parse_package}
\alias{parse_package}
\title{Parse All R Files in a Package}
\usage{
parse_package(path = ".")
}
\arguments{
\item{path}{Path to package root.}
}
\value{
List of all documentation blocks from all R files.
}
\description{
Parse All R Files in a Package
}
\keyword{internal}
