% tinyrox says don't edit this manually, but it can't stop you!
\name{quote_names_in_text}
\alias{quote_names_in_text}
\title{Quote Names in Text}
\usage{
quote_names_in_text(text, names)
}
\arguments{
\item{text}{Text to modify}

\item{names}{Names to quote}
}
\value{
Modified text with names quoted
}
\description{
Wraps unquoted package/software names in single quotes.
}
