% tinyrox says don't edit this manually, but it can't stop you!
\docType{data}
\name{SUPPORTED_DOC_TAGS}
\alias{SUPPORTED_DOC_TAGS}
\title{Supported Documentation Tags}
\format{
An object of class \code{character} of length 22.
}
\usage{
SUPPORTED_DOC_TAGS
}
\description{
Supported Documentation Tags
}
\keyword{internal}
