% tinyrox says don't edit this manually, but it can't stop you!
\docType{data}
\name{SUPPORTED_NS_TAGS}
\alias{SUPPORTED_NS_TAGS}
\title{Supported Namespace Tags}
\format{
An object of class \code{character} of length 6.
}
\usage{
SUPPORTED_NS_TAGS
}
\description{
Supported Namespace Tags
}
\keyword{internal}
