% Generated by rhydrogen: do not edit by hand
% Please edit documentation in R/tags.R
\docType{data}
\name{SUPPORTED_TAGS}
\alias{SUPPORTED_TAGS}
\title{All Supported Tags}
\format{
An object of class \code{character} of length 23.
}
\usage{
SUPPORTED_TAGS
}
\description{
All Supported Tags
}
\keyword{internal}
\keyword{datasets}
