% tinyrox says don't edit this manually, but it can't stop you!
\name{token_root}
\alias{token_root}
\title{Find the Top-Level Expression Containing a Token}
\usage{
token_root(id, parents)
}
\arguments{
\item{id}{Token id from parse data.}

\item{parents}{Named vector mapping token id to parent id.}
}
\value{
Id of the top-level expression.
}
\description{
Find the Top-Level Expression Containing a Token
}
\keyword{internal}
