\name{write_namespace}
\alias{write_namespace}
\title{Write NAMESPACE File}
\description{Write NAMESPACE File}
\usage{
write_namespace(content, path, mode)
}
\arguments{
  \item{content}{NAMESPACE content string.}
  \item{path}{Package root path.}
  \item{mode}{Either "overwrite" or "append".}
}
\keyword{internal}
