% tinyrox says don't edit this manually, but it can't stop you!
\name{write_rd}
\alias{write_rd}
\title{Write Rd File}
\usage{
write_rd(content, name, path = ".")
}
\arguments{
\item{content}{Rd content string.}

\item{name}{Topic name.}

\item{path}{Package root path.}
}
\description{
Write Rd File
}
\keyword{internal}
