% tinyrox says don't edit this manually, but it can't stop you!
\name{check_cran}
\alias{check_cran}
\title{Full CRAN Compliance Check}
\usage{
check_cran(path = ".")
}
\arguments{
\item{path}{Path to package root directory}
}
\value{
List with all issues
}
\description{
Runs all CRAN compliance checks (DESCRIPTION + code).
}
