% Generated by rhydrogen: do not edit by hand
% Please edit documentation in R/document.R
\name{clean}
\alias{clean}
\title{Clean Generated Files}
\usage{
clean(path = ".", namespace = FALSE)
}
\arguments{
\item{path}{Path to package root directory.}

\item{namespace}{Also remove NAMESPACE? Default FALSE.}
}
\description{
Removes all Rd files from man/ directory.
}
\examples{
\dontrun{
clean()
clean(namespace = TRUE)
}
}
