% Generated by rhydrogen: do not edit by hand
% Please edit documentation in R/document.R
\name{clean}
\alias{clean}
\title{Clean Generated Files}
\description{Removes all Rd files from man/ directory.}
\usage{
clean(path = ".", namespace = FALSE)
}
\arguments{
  \item{path}{Path to package root directory.}
  \item{namespace}{Also remove NAMESPACE? Default FALSE.}
}
\examples{
\dontrun{
clean()
clean(namespace = TRUE)
}
}
