% Generated by rhydrogen: do not edit by hand
% Please edit documentation in R/parse.R
\name{extract_formals}
\alias{extract_formals}
\title{Extract Function Formals from Code}
\description{Extract Function Formals from Code}
\usage{
extract_formals(code)
}
\arguments{
  \item{code}{Code starting with "function("}
}
\value{
List with 'names' (argument names) and 'usage' (formatted for Rd).
}
\keyword{internal}
