\name{extract_formals}
\alias{extract_formals}
\title{Extract Function Formals from Code}
\description{Extract Function Formals from Code}
\usage{
extract_formals(code)
}
\arguments{
  \item{code}{Code starting with "function("}
}
\value{
Character vector of formal argument names.
}
\keyword{internal}
