% Generated by rhydrogen: do not edit by hand
% Please edit documentation in R/namespace.R
\name{generate_namespace}
\alias{generate_namespace}
\title{Generate NAMESPACE Content}
\usage{
generate_namespace(blocks)
}
\arguments{
\item{blocks}{List of documentation blocks from parse_package().}
}
\value{
Character string of NAMESPACE content.
}
\description{
Generate NAMESPACE Content
}
\keyword{internal}
