% tinyrox says don't edit this manually, but it can't stop you!
\name{get_package_name}
\alias{get_package_name}
\title{Get Package Name from DESCRIPTION}
\usage{
get_package_name(path)
}
\arguments{
\item{path}{Package root path.}
}
\value{
Package name.
}
\description{
Get Package Name from DESCRIPTION
}
\keyword{internal}
