% Generated by rhydrogen: do not edit by hand
% Please edit documentation in R/namespace.R
\docType{data}
\name{KNOWN_S3_GENERICS}
\alias{KNOWN_S3_GENERICS}
\title{Known S3 Generic Functions}
\format{
An object of class \code{character} of length 122.
}
\usage{
KNOWN_S3_GENERICS
}
\description{
List of base R S3 generics for auto-detection when @export is used.
}
\keyword{internal}
