% Generated by rhydrogen: do not edit by hand
% Please edit documentation in R/parse.R
\name{parse_formals_text}
\alias{parse_formals_text}
\title{Parse Formals Text}
\usage{
parse_formals_text(text)
}
\arguments{
\item{text}{Text containing function arguments.}
}
\value{
List with 'names' (argument names) and 'usage' (formatted for Rd).
}
\description{
Parse Formals Text
}
\keyword{internal}
