% Generated by rhydrogen: do not edit by hand
% Please edit documentation in R/parse.R
\name{parse_package}
\alias{parse_package}
\title{Parse All R Files in a Package}
\description{Parse All R Files in a Package}
\usage{
parse_package(path = ".")
}
\arguments{
  \item{path}{Path to package root.}
}
\value{
List of all documentation blocks from all R files.
}
\keyword{internal}
