% Generated by rhydrogen: do not edit by hand
% Please edit documentation in R/tags.R
\docType{data}
\name{SUPPORTED_DOC_TAGS}
\alias{SUPPORTED_DOC_TAGS}
\title{Supported Documentation Tags}
\format{
An object of class \code{character} of length 18.
}
\usage{
SUPPORTED_DOC_TAGS
}
\description{
Supported Documentation Tags
}
\keyword{internal}
