% Generated by rhydrogen: do not edit by hand
% Please edit documentation in R/tags.R
\docType{data}
\name{SUPPORTED_NS_TAGS}
\alias{SUPPORTED_NS_TAGS}
\title{Supported Namespace Tags}
\format{
An object of class \code{character} of length 5.
}
\usage{
SUPPORTED_NS_TAGS
}
\description{
Supported Namespace Tags
}
\keyword{internal}
