% Generated by rhydrogen: do not edit by hand
% Please edit documentation in R/namespace.R
\name{write_namespace}
\alias{write_namespace}
\title{Write NAMESPACE File}
\usage{
write_namespace(content, path = ".", mode = "overwrite")
}
\arguments{
\item{content}{NAMESPACE content string.}

\item{path}{Package root path.}

\item{mode}{Either "overwrite" or "append".}
}
\description{
Write NAMESPACE File
}
\keyword{internal}
