% Generated by rhydrogen: do not edit by hand
% Please edit documentation in R/rd.R
\name{write_rd}
\alias{write_rd}
\title{Write Rd File}
\usage{
write_rd(content, name, path = ".")
}
\arguments{
\item{content}{Rd content string.}

\item{name}{Topic name.}

\item{path}{Package root path.}
}
\description{
Write Rd File
}
\keyword{internal}
